tapapi/requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
requests==2.27.1
urllib3==1.26.8
Werkzeug==2.0.2